Use a variable for data.train.next_batch()
Instead of an arbitrary constant, I think it might be better to document it to use a variable such as `BATCH_SIZE`. I can't recall which forum and question it was, but a user (someone who was still learning TensorFlow) was confused because of the loop counter in [Reading Data](https://www.tensorflow.org/api_guides/python/reading_data#csv_files) (CSV file section), `for i in range(1200)`. The user thought that 1200 was required constant or something.